updated files and proxy
