+++ client/package.json
@@ -1,6 +1,7 @@
 {
   "name": "client",
-  "version": "0.1.0",
+	"version": "0.1.0",
+	"proxy": "http://localhost:3001",
   "private": true,
   "dependencies": {
     "react": "^16.7.0",
